requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.10.4
feedparser==6.0.11
pytz==2024.1